pyproject.toml
[project]
name = "papermind"
version = "0.1.0"
description = "Local AI research system: RAG, knowledge graphs, and paper-to-code generation on 12GB VRAM"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
    "chromadb>=0.6.0",
    "sentence-transformers>=3.4.0",
    "faiss-cpu>=1.9.0",
    "pymupdf>=1.25.0",
    "networkx>=3.4",
    "tiktoken>=0.9.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "click>=8.1.0",
    "einops>=0.8.2",
    "datasketch>=1.9.0",
    "grobid-client-python>=0.1.4",
    "mineru>=2.7.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.120.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]
streamlit = ["streamlit>=1.42.0", "pyvis>=0.3.2"]
agents = ["langgraph>=0.3.0", "langchain-core>=0.3.0"]
local-llm = [
    "bitsandbytes>=0.45.0",
    "accelerate>=1.3.0",
]
sandbox = ["docker>=7.1.0"]
math = ["latex2sympy2>=1.9.1", "sympy>=1.13.0"]
data = ["datasets>=3.2.0"]

[project.scripts]
papermind = "papermind.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/papermind"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
